requirements.txt
chardet
Django==1.6.9
MySQL-python
djangorestframework==3.0.3
gunicorn
gevent
celery
netaddr
django-filter
kazoo
redis
pika
GitPython
django-redis
elasticsearch==5.2
requests==2.10.0
pychecker==0.8.19
PyYAML==3.11
xlwt==1.1.2
Pillow==3.4.2
django-auth-ldap==1.2.8
python-ldap==2.4.25
python-hessian==1.0.2
django-debug-toolbar==1.3.2
sqlparse==0.1.19
